feat: add User custom decorator

+++ src/todos/todos.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
 import { UpdateTodoDto } from './dto/update-todo.dto';
 
@@ -13,23 +17,41 @@ export class TodosService {
   }
 
   async findOneTodo(id: number, userId: number) {
-    return await this.prisma.todo.findFirst({ where: { id, userId } });
+    const todo = await this.prisma.todo.findFirst({ where: { id } });
+    if (!todo) {
+      throw new NotFoundException('todo with provided id is not found');
+    }
+    if (todo.userId !== userId) {
+      throw new ForbiddenException('You are not permitted to access this todo');
+    }
+    return todo;
   }
 
-  async listTodos(userId: number) {
-    return await this.prisma.todo.findMany({ where: { userId } });
-  }
-
-  async updateTodo(id: number, payload: UpdateTodoDto) {
-    return await this.prisma.todo.update({
-      where: {
-        id,
-      },
-      data: payload,
+  async listTodos(userId: number, limit: number = 5, offset: number = 0) {
+    return await this.prisma.todo.findMany({
+      skip: offset,
+      take: limit,
+      where: { userId },
     });
   }
 
-  async deleteTodo(id: number) {
-    return await this.prisma.todo.delete({ where: { id } });
+  async updateTodo(id: number, userId: number, payload: UpdateTodoDto) {
+    const todo = await this.findOneTodo(id, userId);
+    if (todo.id) {
+      return await this.prisma.todo.update({
+        where: {
+          id,
+          userId,
+        },
+        data: payload,
+      });
+    }
+  }
+
+  async deleteTodo(id: number, userId: number) {
+    const todo = await this.findOneTodo(id, userId);
+    if (todo.id) {
+      return await this.prisma.todo.delete({ where: { id, userId } });
+    }
   }
 }
